rename useDocumentTitle
