+++ src/features/apiDelegation/offered/OverviewPage/OverviewPage.tsx
@@ -4,10 +4,10 @@ import * as React from 'react';
 import { Page, PageHeader, PageContent, PageContainer } from '@/components';
 import ApiIcon from '@/assets/Api.svg?react';
 import { useMediaQuery } from '@/resources/hooks';
+import { useDocumentTitle } from '@/resources/hooks/useDocumentTitle';
 
 import { OverviewPageContent } from '../../components/OverviewPageContent';
 import { LayoutState } from '../../components/LayoutState';
-import { useDocumentTitle } from '@/resources/utils/pageUtils';
 
 export const OverviewPage = () => {
   const { t } = useTranslation('common');
