+++ src/features/apiDelegation/received/OverviewPage/OverviewPage.tsx
@@ -4,7 +4,7 @@ import * as React from 'react';
 import { Page, PageHeader, PageContent, PageContainer } from '@/components';
 import ApiIcon from '@/assets/Api.svg?react';
 import { useMediaQuery } from '@/resources/hooks';
-import { useDocumentTitle } from '@/resources/utils/pageUtils';
+import { useDocumentTitle } from '@/resources/hooks/useDocumentTitle';
 
 import { OverviewPageContent } from '../../components/OverviewPageContent';
 import { LayoutState } from '../../components/LayoutState';
