+++ src/features/singleRight/delegate/ReceiptPage/ReceiptPage.tsx
@@ -12,10 +12,10 @@ import { useAppDispatch, useAppSelector } from '@/rtk/app/hooks';
 import { resetServicesWithStatus } from '@/rtk/features/singleRights/singleRightsSlice';
 import { GroupElements } from '@/components/GroupElements/GroupElements';
 import { redirectToSevicesAvailableForUser } from '@/resources/utils';
+import { useDocumentTitle } from '@/resources/hooks/useDocumentTitle';
 
 import classes from './ReceiptPage.module.css';
 import { ActionBarSection } from './ActionBarSection/ActionBarSection';
-import { useDocumentTitle } from '@/resources/utils/pageUtils';
 
 export const ReceiptPage = () => {
   const { t } = useTranslation();
