+++ src/resources/hooks/useDocumentTitle.tsx
